cleaning up display method for view controllers
