+++ TranquilApp/Pages/ExerciseModelData.swift
@@ -1,0 +1,49 @@
+//
+//  ExerciseModelData.swift
+//  TranquilApp
+//
+//  Created by Heather Dinh on 1/30/23.
+//
+
+
+import Foundation
+import Combine
+
+final class ExerciseModelData: ObservableObject {
+    @Published var landmarks: [Landmark] = load("exercisesData.json")
+//    var hikes: [Hike] = load("hikeData.json")
+//    @Published var profile = Profile.default
+
+    var features: [Landmark] {
+        landmarks.filter { $0.isFeatured }
+    }
+
+    var categories: [String: [Landmark]] {
+        Dictionary(
+            grouping: landmarks,
+            by: { $0.category.rawValue }
+        )
+    }
+}
+
+func load<T: Decodable>(_ filename: String) -> T {
+    let data: Data
+
+    guard let file = Bundle.main.url(forResource: filename, withExtension: nil)
+        else {
+            fatalError("Couldn't find \(filename) in main bundle.")
+    }
+
+    do {
+        data = try Data(contentsOf: file)
+    } catch {
+        fatalError("Couldn't load \(filename) from main bundle:\n\(error)")
+    }
+
+    do {
+        let decoder = JSONDecoder()
+        return try decoder.decode(T.self, from: data)
+    } catch {
+        fatalError("Couldn't parse \(filename) as \(T.self):\n\(error)")
+    }
+}
