+++ TranquilApp/Pages/PageView.swift
@@ -27,7 +27,7 @@ struct PageView<Page: View>: View {
 struct PageView_Previews: PreviewProvider {
     static var previews: some View {
         PageView(
-            pages: ModelData().features.map { FeatureCard(landmark: $0) }
+            pages: ExerciseModelData().features.map { FeatureCard(landmark: $0) }
         )
             .aspectRatio(3 / 2, contentMode: .fit)
     }
