+++ TranquilApp/Views/HistoryView.swift
@@ -22,11 +22,6 @@ struct HistoryView: View {
                 .font(.system(size: 60, weight: .heavy))
                 .bold()
                 .foregroundColor(Color.teal)
-            Button(action: {
-                self.presentationMode.wrappedValue.dismiss()
-            }) {
-                Text("Dismiss")
-            }
             Spacer()
             BarChart()
                 .previewLayout(.sizeThatFits)
